Add failing test case for newlines
There are two modes of failure here. The first added test string shows
that indentation is not properly added after a newline. The second
part, which is a new type of test, shows that the line following the
newline does not extend to the full length limit before it is broken
again by this module. (It also shows the first issue, but not part of
the existing test framework, so I thought maybe both are valuable.) To
illustrate:

Expected (there's a newline just before the capital B):

    A a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a
        a a a a a a a a a a a
        B b b b b b b b b b b b b b b b b b b b b b b b b b b b b b b b b b b b b b
        b b b b b b b b b b b

Actual:

    A a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a a
        a a a a a a a a a a a
    B b b b b b b b b b b b b b b b b b b b b b b b b b b
        b b b b b b b b b b b b b b b b b b b b b b b b